+++ Calculator/src/main/java/application/calculator/StandardCalc.java
@@ -79,11 +79,11 @@ public class StandardCalc implements Calculator {
 
         return rpCalc.evaluate(postfixExpr);
 
-      } catch (Exception ex) {
-        throw new InvalidExpression();
+      } catch (Exception exception) {   
+        throw new InvalidExpression(exception.getMessage());
       }
     }
-    throw new InvalidExpression();
+    throw new InvalidExpression("The expression seems to be in postfix notation!");
   }
 
   /**
@@ -126,5 +126,10 @@ public class StandardCalc implements Calculator {
       default:
         return Symbol.INVALID;
     }
+  }
+  
+  @Override
+  public String toString() {
+    return "Standard (infix) calculator";
   }
 }