+++ Calculator/src/main/java/application/view/AsciiView.java
@@ -14,6 +14,12 @@ public class AsciiView implements ViewInterface {
   // The current question that the calculator must solve: entered like ? 3 * ( 5 + 4 )
   private String question;
 
+  // The current type of calculator being used - a notification given to the user when they first
+  // run the program & change type
+  private String typeNotification;
+
+  private String history;
+
   // This method will be injected so we can ask the controller to calculate
   Runnable doCalculation;
 
@@ -23,15 +29,21 @@ public class AsciiView implements ViewInterface {
   private void menu() {
     Scanner s = new Scanner(System.in);
     boolean finished = false;
+    System.out.println(typeNotification);
     help();
     while (!finished && s.hasNext()) {
       String t = s.next();
       switch (t.toUpperCase().charAt(0)) {
-        case 'C': // Ask the controller to calculate
+        case '?': // Set and calculate current question
+          question = s.nextLine();
           doCalculation.run();
           break;
-        case '?': // Set current question
-          question = s.nextLine();
+        case 'C': // Change the notation type to the one currently not being used
+          setCalculatorType.accept("");
+          System.out.println(typeNotification);
+          break;
+        case 'H':
+          System.out.println("\n\nHistory:\n" + history + "\n");
           break;
         case 'Q':
           System.out.println("Bye");
@@ -46,10 +58,9 @@ public class AsciiView implements ViewInterface {
 
   private void help() {
     System.out.println("Use one of the following:");
-    System.out.println("  ? Expression - to set expression");
-    System.out.println("  C - to calculate");
-    System.out.println("  S - change to a standard calculator");
-    System.out.println("  R - change to a reverse polish calculator");
+    System.out.println("  ? Expression - to calculate an expression");
+    System.out.println("  C - to change notation type");
+    System.out.println("  H - to display history");
     System.out.println("  Q - to exit");
   }
 
@@ -61,11 +72,11 @@ public class AsciiView implements ViewInterface {
   @Override
   public void setAnswer(String answer) {
     System.out.println(answer);
-  }
 
-  @Override
-  public void addTypeObserver(Consumer<String> c) {
-
+    if (answer.contains("invalid")) {
+      answer = "invalid expression"; // less clutter in history view
+    }
+    history += question + " = " + answer + "\n";
   }
 
   @Override
@@ -73,32 +84,46 @@ public class AsciiView implements ViewInterface {
     menu();
   }
 
+  // Methods for registering an observer and for accessing user data in the UI
+  // These methods build the Observer/Observable pattern
+
   @Override
   public void addCalculateObserver(Runnable function) {
     doCalculation = function;
   }
-  
-  /////////////////////////////////////////////////////////////////////////////////  
+
+  @Override
+  public void addTypeObserver(Consumer<String> c) {
+    setCalculatorType = c;
+  }
+
+  @Override
+  public void setCalculatorNotification(String notification) {
+    typeNotification = notification;
+  }
+
+  /////////////////////////////////////////////////////////////////////////////////
   // Block for creating an instance variable for others to use - singleton.
-  
+
   /**
    * A private constructor for the singleton View. Constructs the view.
    */
   private AsciiView() {
     doCalculation = null;
-    setCalculatorType = null; 
+    setCalculatorType = null;
+    history = "";
   }
-  
+
   private static volatile AsciiView instance = new AsciiView();
 
   /**
    * This is a Singleton View made available through this method.
    * 
-   * @return the single object representing this view
+   * @return The single object representing this view.
    */
   public static synchronized AsciiView getInstance() {
     if (instance == null) {
-      instance = new AsciiView(); 
+      instance = new AsciiView();
     }
     return instance;
   }